requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dispy==4.15.2
Django==4.2.10
django-celery-results==2.5.1
django-clearcache==1.2.1
django-cors-headers==4.3.1
django-db-logger==0.1.13
django-post-request-task==0.5
django-redis==5.4.0
django-truncate==0.1
djangomix==1.1.3
djangorestframework==3.14.0
feedparser==6.0.11
greenlet==3.0.3
idna==3.6
kombu==5.3.4
mysqlclient==2.2.4
oauthlib==3.2.2
prompt-toolkit==3.0.41
pycos==4.12.2
pymemcache==4.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
six==1.16.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
